+++ jest.config.ts
@@ -1,0 +1,15 @@
+import type { Config } from '@jest/types';
+
+const config: Config.InitialOptions = {
+  verbose: true,
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  roots: ['./tests'],
+  testMatch: ['**/__tests__**/*.ts', '**/?(*.)+(spec|test).ts'],
+  transform: {
+    '^.+\\.ts$': 'ts-jest',
+  },
+  setupFilesAfterEnv: ['./tests/setup.ts']
+};
+
+export default config;
