+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "NODE_ENV=test jest --config jest.config.ts",
+    "test:watch": "NODE_ENV=test jest --config jest.config.ts --watch",
     "start": "npx nodemon index.ts"
   },
   "author": "",
@@ -12,7 +13,6 @@
   "dependencies": {
     "@types/express": "^5.0.0",
     "@types/node": "^22.13.4",
-    "@types/sqlite3": "^5.1.0",
     "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
     "create-react-app": "^5.1.0",
@@ -28,5 +28,13 @@
     "tiny-secp256k1": "^2.2.3",
     "ts-node": "^10.9.2",
     "uuid": "^11.1.0"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.5.14",
+    "@types/sqlite3": "^5.1.0",
+    "@types/supertest": "^6.0.2",
+    "jest": "^29.7.0",
+    "supertest": "^7.0.0",
+    "ts-jest": "^29.2.6"
   }
 }
